Adjusted styles :nail_care:
